requirements.txt
streamlit>=1.40.0
langchain==1.0.3
langchain-core==1.0.2
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-ollama==1.0.0
langchain-huggingface==1.0.0
langchain-pinecone==0.0.1
chromadb>=0.5.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
ollama>=0.2.0
numpy==1.26.4
langchain-openai>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
pinecone-client>=3.2.0
